fix(pi05): green CI (ruff + strict mypy) and translate scoping doc to English
CI was red on `ruff check src tests` (10 lint errors) and would have failed
`mypy src/odyssey` next (26 strict errors in the new pi05 modules).

- ruff: sort test imports (I001), drop redundant int(round()) casts (RUF046),
  un-quote NormStats return annotations (UP037), replace ambiguous `×` in a
  docstring (RUF002), and collapse the nested COMPLETION branch in
  PlannedEvalRuntime.check_advance (SIM102).
- mypy: exempt the numpy-only openpi mirrors (pi05_norm/fast/transforms + the
  openpi_client wrapper pi05.py) from strict checking via a tool.mypy override,
  exactly as the GR00T transforms/eval scripts already are — their behaviour is
  pinned by tests/unit/test_pi05_*.py. Fix the one general-module error properly:
  cast the detector callable in completion_gate._as_callable (no-any-return).
- docs: translate docs/pi05-scoping.md to English and update the three
  in-code section cross-references that pointed at the old Spanish headings.

ruff, mypy --strict, and pytest (486 passed, 5 skipped) all green locally.

+++ tests/unit/test_pi05_fast.py
@@ -39,7 +39,6 @@ from odyssey.runners.evals.pi05_fast import (
     unnormalize_from_unit,
 )
 
-
 # ---------------------------------------------------------------------------
 # Fake FAST processor — a dependency-free stand-in for the HF AutoProcessor.
 #
@@ -64,7 +63,7 @@ class _FakeFastProcessor:
         self.encode_calls.append(arr)
         out: list[list[int]] = []
         for item in arr:  # (T, D)
-            out.append([int(round(v * self.SCALE)) + self.OFFSET
+            out.append([round(v * self.SCALE) + self.OFFSET
                         for v in item.reshape(-1)])
         return out
 
@@ -124,7 +123,7 @@ def test_pi05_fast_decode_of_synthetic_tokens() -> None:
     # Hand-build tokens for a known (2, 3) chunk under the fake's fixed-point scheme.
     values = np.array([[0.0, 0.5, -0.5], [0.25, -0.25, 1.0]], dtype=np.float64)
     P = _FakeFastProcessor
-    tokens = [int(round(v * P.SCALE)) + P.OFFSET for v in values.reshape(-1)]
+    tokens = [round(v * P.SCALE) + P.OFFSET for v in values.reshape(-1)]
 
     chunk = codec.decode(tokens, time_horizon=2, action_dim=3)
